Move hello world tests out into separate test file
