+++ tests/x86_64_emulator.rs
@@ -1,313 +1,15 @@
 mod common;
 
-use std::{path::Path, rc::Rc};
-
 use common::{Memory, TracingEmulator, x86_64_sleigh};
-use libsla::{Address, GhidraSleigh, Sleigh, VarnodeData};
-use pcode_ops::{PcodeOps, convert::PcodeValue};
-use sym::{self, Evaluator, SymbolicBit, SymbolicBitVec, SymbolicByte, VariableAssignments};
+use libsla::{Address, Sleigh, VarnodeData};
+use pcode_ops::PcodeOps;
+use sym::{self, Evaluator, SymbolicBitVec, VariableAssignments};
 use symbolic_pcode::{
-    arch::{
-        self,
-        x86::{emulator::EmulatorX86, processor::ProcessorHandlerX86},
-    },
+    arch::{self, x86::processor::ProcessorHandlerX86},
     emulator::StandardPcodeEmulator,
-    kernel::linux::{LinuxArchConfig, LinuxKernel},
     mem::{MemoryTree, VarnodeDataStore},
     processor::{self, BranchingProcessor, Processor, ProcessorState},
 };
-
-const INITIAL_STACK: u64 = 0x8000000000;
-const EXIT_IP_ADDR: u64 = 0xFEEDBEEF0BADF00D;
-
-fn initialize_libc_stack(memory: &mut Memory, sleigh: &impl Sleigh) {
-    // The stack for libc programs:
-    // * argc
-    // * argv - list must be terminated by NULL pointer
-    // * envp - list must be terminated by NULL pointer
-    // * auxv - list must be terminated by NULL pointer
-    let ram = sleigh
-        .address_space_by_name("ram")
-        .expect("failed to find ram");
-    let argc = VarnodeData {
-        address: Address {
-            offset: INITIAL_STACK,
-            address_space: ram.clone(),
-        },
-        size: 8,
-    };
-    memory
-        .write(&argc, SymbolicBitVec::constant(1, u64::BITS as usize))
-        .expect("failed to initialize argc on stack");
-
-    // The argv list must be terminated by null pointer. Setting program name to null AND
-    // terminating the list with NULL, whence 16 bytes
-    //
-    // MUSL has support for null program name:
-    // https://git.musl-libc.org/cgit/musl/tree/src/env/__libc_start_main.c
-    let argv = VarnodeData {
-        address: Address {
-            offset: argc.address.offset + argc.size as u64,
-            address_space: ram.clone(),
-        },
-        size: 16,
-    };
-    memory
-        .write(&argv, SymbolicBitVec::constant(0, (2 * u64::BITS) as usize))
-        .expect("failed to initialize argv");
-
-    let envp = VarnodeData {
-        address: Address {
-            offset: argv.address.offset + argv.size as u64,
-            address_space: ram.clone(),
-        },
-        size: 8,
-    };
-    memory
-        .write(&envp, SymbolicBitVec::constant(0, u64::BITS as usize))
-        .expect("failed to initialize envp");
-
-    // MUSL targets initialize the libc pagesize using aux[AT_PAGESZ]. For architectures without a
-    // hardcoded value the libc pagesize is used. For example x86-64 has a hardcoded value of 4096
-    // and so this is ignored. However for aarch64 there is no hardcoded value. This must be
-    // supplied otherwise a pagesize of 0 will be used.
-    let mut auxv = VarnodeData {
-        address: Address {
-            offset: envp.address.offset + envp.size as u64,
-            address_space: ram.clone(),
-        },
-        size: 8,
-    };
-
-    // The index for AT_PAGESZ
-    let at_pagesz = 6;
-    memory
-        .write(
-            &auxv,
-            SymbolicBitVec::constant(at_pagesz, u64::BITS as usize),
-        )
-        .expect("failed to write AT_PAGESZ into auxv");
-    auxv.address.offset += auxv.size as u64;
-
-    let page_size = 4096;
-    memory
-        .write(
-            &auxv,
-            SymbolicBitVec::constant(page_size, u64::BITS as usize),
-        )
-        .expect("failed to write page size into auxv");
-    auxv.address.offset += auxv.size as u64;
-
-    memory
-        .write(&auxv, SymbolicBitVec::constant(0, u64::BITS as usize))
-        .expect("failed to initialize auxv");
-}
-
-fn memory_with_image(
-    sleigh: &GhidraSleigh,
-    image: impl AsRef<Path>,
-    pc_register: &VarnodeData,
-) -> Memory {
-    use elf::ElfBytes;
-    use elf::abi::PT_LOAD;
-    use elf::endian::AnyEndian;
-
-    let mut memory = Memory::default();
-
-    // Write image into memory
-    let data = std::fs::read(image).expect("failed to read image file");
-
-    let elf = ElfBytes::<AnyEndian>::minimal_parse(&data).expect("failed to parse elf");
-    for segment in elf.segments().expect("failed to get segments") {
-        if segment.p_type != PT_LOAD {
-            // Not a loadable segment
-            continue;
-        }
-
-        let data_location = VarnodeData {
-            address: Address {
-                offset: segment.p_vaddr,
-                address_space: sleigh.default_code_space(),
-            },
-
-            // The initial set of data is from the file
-            size: segment.p_filesz as usize,
-        };
-
-        // Write the segment from the file into memory
-        let offset = segment.p_offset as usize;
-        let file_size = segment.p_filesz as usize;
-        memory
-            .write(
-                &data_location,
-                data[offset..offset + file_size].iter().copied().collect(),
-            )
-            .expect("failed to write image section into memory");
-
-        println!(
-            "Loaded segment {vaddr:#x} from {offset:#x} ({file_size})",
-            vaddr = segment.p_vaddr
-        );
-
-        // If the virtual size is larger than the file size then zero out the remainder
-        let zero_size = (segment.p_memsz - segment.p_filesz) as usize;
-        if zero_size > 0 {
-            let zeros = vec![0u8; zero_size];
-            let zeros_location = VarnodeData {
-                address: Address {
-                    offset: segment.p_vaddr + segment.p_filesz,
-                    address_space: sleigh.default_code_space(),
-                },
-                size: zero_size,
-            };
-            println!(
-                "Zeroing segment {vaddr:#x} at {zeros_location}",
-                vaddr = segment.p_vaddr
-            );
-            memory
-                .write(&zeros_location, zeros.into_iter().collect())
-                .expect("failed to write zeros into memory");
-        }
-    }
-
-    // Init RIP to entry
-    memory
-        .write(pc_register, elf.ehdr.e_entry.into())
-        .expect("failed to initialize PC register");
-
-    memory
-}
-
-fn init_registers_x86_64(sleigh: &impl Sleigh, memory: &mut Memory) {
-    let rsp = sleigh.register_from_name("RSP").expect("invalid register");
-    let registers = ["RAX", "RBX", "RCX", "RDX", "RSI", "RDI", "RBP"]
-        .into_iter()
-        .map(str::to_owned)
-        .chain((8..16).map(|n| format!("R{n}")));
-
-    init_registers(sleigh, memory, &rsp, registers);
-
-    // Initialize the DF register to 0. It appears to be a convention that whenever STD is called
-    // CLD is called thereafter to reset it. There is a bug (?) in MUSL where REP STOS is used
-    // without ensuring the flag is cleared
-
-    // 0000000000449f4d <__init_libc>:
-    //   449f4d:       53                      push   rbx
-    //   449f4e:       48 89 fa                mov    rdx,rdi
-    //   449f51:       31 c0                   xor    eax,eax
-    //   449f53:       b9 4c 00 00 00          mov    ecx,0x4c
-    //   449f58:       48 81 ec 50 01 00 00    sub    rsp,0x150
-    //   449f5f:       48 8d 7c 24 20          lea    rdi,[rsp+0x20]
-    //   449f64:       f3 ab                   rep stos DWORD PTR es:[rdi],eax
-
-    let df_register = sleigh
-        .register_from_name("DF")
-        .expect("failed to get DF register");
-    memory
-        .write(&df_register, 0u8.into())
-        .expect("failed to init DF register");
-}
-
-fn init_registers_aarch64(sleigh: &impl Sleigh, memory: &mut Memory) {
-    let sp = sleigh
-        .register_from_name("sp")
-        .expect("invalid stack register");
-    let registers = (0..30).map(|n| format!("x{n}"));
-
-    init_registers(sleigh, memory, &sp, registers);
-
-    // Init link register value to final return address
-    let link_register = sleigh
-        .register_from_name("x30")
-        .expect("invalid link register");
-    memory
-        .write(&link_register, EXIT_IP_ADDR.into())
-        .expect("failed to initialize link register");
-
-    // Initialize system register `dczid_el0`
-    //
-    // Per DDI0487L_b_a-profile_architecture_reference_manual.pdf
-    //
-    // Indicates the block size that is written with byte values of 0 by the DC ZVA (Data Cache
-    // Zero by Address) System instruction.
-    //
-    // - Bits [63:5] Reserved, RES0.
-    // - Bit [4] Data Zero Prohibited (DZP). This field indicates whether the use of `DC ZVA`
-    // instruction is permitted (0b0) or prohibited (0b1).
-    // - Bits [3:0] Block Size (BS). Log2 of the block size in words.
-    let dczid_el0 = sleigh
-        .register_from_name("dczid_el0")
-        .expect("unknown register");
-    memory
-        .write(&dczid_el0, 0x10u64.into())
-        .expect("failed to initialize dczid_el0");
-}
-
-fn init_registers(
-    sleigh: &impl Sleigh,
-    memory: &mut Memory,
-    stack_register: &VarnodeData,
-    registers: impl IntoIterator<Item = String>,
-) {
-    let mut bitvar = 0;
-    for register_name in registers.into_iter() {
-        let mut bytes = Vec::with_capacity(8);
-        for _ in 0..8 {
-            let byte: SymbolicByte = [
-                SymbolicBit::Variable(bitvar),
-                SymbolicBit::Variable(bitvar + 1),
-                SymbolicBit::Variable(bitvar + 2),
-                SymbolicBit::Variable(bitvar + 3),
-                SymbolicBit::Variable(bitvar + 4),
-                SymbolicBit::Variable(bitvar + 5),
-                SymbolicBit::Variable(bitvar + 6),
-                SymbolicBit::Variable(bitvar + 7),
-            ]
-            .into();
-            bytes.push(byte);
-            bitvar += 8;
-        }
-
-        let register = sleigh
-            .register_from_name(&register_name)
-            .unwrap_or_else(|err| panic!("invalid register {register_name}: {err}"));
-        memory
-            .write(&register, bytes.into_iter().collect())
-            .unwrap_or_else(|err| panic!("failed to write register {register_name}: {err}"));
-    }
-
-    // Init stack register to stack address
-    memory
-        .write(stack_register, INITIAL_STACK.into())
-        .expect("failed to initialize stack register");
-}
-
-fn find_test_fixture(
-    messages: escargot::CommandMessages,
-) -> Result<std::path::PathBuf, Vec<escargot::error::CargoResult<escargot::Message>>> {
-    let messages: Vec<_> = messages.into_iter().collect();
-    let executable_path = messages.iter().find_map(|result| {
-        result
-            .as_ref()
-            .ok()
-            .into_iter()
-            .filter_map(|msg| {
-                msg.decode()
-                    .ok()
-                    .into_iter()
-                    .filter_map(|msg| match msg {
-                        escargot::format::Message::CompilerArtifact(artifact) => {
-                            artifact.executable.map(|p| p.into_owned())
-                        }
-                        _ => None,
-                    })
-                    .next()
-            })
-            .next()
-    });
-
-    executable_path.ok_or(messages)
-}
 
 /// Confirms the functionality of general-purpose x86-64 registers and overlapping behavior.
 #[test]
@@ -521,153 +223,6 @@ fn doubler_32b() -> processor::Result<()> {
 }
 
 #[test]
-fn hello_world_x86_linux() -> processor::Result<()> {
-    flexi_logger::Logger::try_with_env()
-        .unwrap()
-        .start()
-        .unwrap();
-
-    // Build test fixture first
-    let messages = escargot::CargoBuild::new()
-        .bin("linux-syscalls")
-        .manifest_path("./test-fixtures/linux-syscalls/Cargo.toml")
-        .target("x86_64-unknown-linux-musl")
-        .env(
-            "RUSTFLAGS",
-            "-Ctarget-feature=+crt-static -Crelocation-model=static",
-        )
-        .exec()
-        .unwrap();
-
-    let image = match find_test_fixture(messages) {
-        Ok(image) => image,
-        Err(messages) => {
-            panic!("Failed to find test fixture: {messages:?}");
-        }
-    };
-
-    let sleigh = Rc::new(x86_64_sleigh().expect("failed to build sleigh"));
-    let rip = sleigh.register_from_name("RIP").expect("invalid register");
-    let mut memory = memory_with_image(&sleigh, image, &rip);
-    init_registers_x86_64(sleigh.as_ref(), &mut memory);
-    initialize_libc_stack(&mut memory, sleigh.as_ref());
-
-    let handler = ProcessorHandlerX86::new(sleigh.as_ref());
-    let emulator = EmulatorX86::with_kernel(sleigh.clone(), LinuxKernel::default());
-    let mut processor = Processor::new(memory, emulator, handler);
-
-    loop {
-        processor.step(sleigh.as_ref())?;
-
-        // Debug
-        if matches!(processor.state(), ProcessorState::Decode(_)) {
-            let disassembly = processor.disassemble(sleigh.as_ref())?;
-            let encoded_instr = processor
-                .memory()
-                .read(&disassembly.origin)?
-                .into_le_bytes()
-                .map(|byte| u8::try_from(byte).map_or("xx".to_string(), |b| format!("{b:02x}")))
-                .collect::<Vec<_>>()
-                .join(" ");
-
-            println!("Encoded instruction from memory: {encoded_instr}");
-            println!("Decoded: {disassembly}");
-        }
-
-        if matches!(processor.state(), ProcessorState::Halt) {
-            assert_eq!(
-                processor.emulator().kernel().exit_status(),
-                Some(0),
-                "exit code should be 0"
-            );
-            return Ok(());
-        }
-    }
-}
-
-#[test]
-fn hello_world_aarch64_linux() -> processor::Result<()> {
-    flexi_logger::Logger::try_with_env()
-        .unwrap()
-        .start()
-        .unwrap();
-
-    // Build test fixture first
-    let messages = escargot::CargoBuild::new()
-        .bin("linux-syscalls")
-        .manifest_path("./test-fixtures/linux-syscalls/Cargo.toml")
-        .target("aarch64-unknown-linux-musl")
-        .env(
-            "RUSTFLAGS",
-            "-Ctarget-feature=+crt-static -Crelocation-model=static",
-        )
-        // On Ubuntu can install with sudo apt install gcc-aarch64-linux-gnu
-        .env(
-            "CARGO_TARGET_AARCH64_UNKNOWN_LINUX_MUSL_LINKER",
-            "aarch64-linux-gnu-gcc",
-        )
-        .exec()
-        .unwrap();
-
-    let image = match find_test_fixture(messages) {
-        Ok(image) => image,
-        Err(messages) => {
-            panic!("Failed to find test fixture: {messages:?}");
-        }
-    };
-
-    let sleigh = Rc::new(common::aarch64_sleigh().expect("failed to build sleigh"));
-    let pc = sleigh.register_from_name("pc").expect("invalid register");
-    let mut memory = memory_with_image(&sleigh, image, &pc);
-    init_registers_aarch64(sleigh.as_ref(), &mut memory);
-    initialize_libc_stack(&mut memory, sleigh.as_ref());
-
-    let handler = arch::aarch64::processor::ProcessorHandler::new(sleigh.as_ref());
-    let emulator = arch::aarch64::emulator::Emulator::with_kernel(
-        sleigh.clone(),
-        LinuxKernel::with_config(arch::aarch64::linux::config()),
-    );
-    let mut processor = Processor::new(memory, emulator, handler);
-
-    loop {
-        let print_pcode = matches!(processor.state(), ProcessorState::Decode(_));
-        processor.step(sleigh.as_ref())?;
-
-        if print_pcode {
-            if let ProcessorState::Execute(e) = processor.state() {
-                for instr in &e.pcode().instructions {
-                    println!("{instr}");
-                }
-            }
-        }
-
-        // Debug
-        if matches!(processor.state(), ProcessorState::Decode(_)) {
-            let disassembly = processor.disassemble(sleigh.as_ref())?;
-            let encoded_instr = processor
-                .memory()
-                .read(&disassembly.origin)?
-                .into_le_bytes()
-                .map(|byte| u8::try_from(byte).map_or("xx".to_string(), |b| format!("{b:02x}")))
-                .collect::<Vec<_>>()
-                .join(" ");
-
-            println!("Encoded instruction from memory: {encoded_instr}");
-            println!("Decoded: {instr}", instr = disassembly.instructions[0]);
-        }
-
-        if matches!(processor.state(), ProcessorState::Halt) {
-            assert_eq!(
-                processor.emulator().kernel().exit_status(),
-                Some(0),
-                "exit code should be 0"
-            );
-            return Ok(());
-        }
-    }
-}
-
-#[test]
 fn pcode_coverage() -> processor::Result<()> {
     // Build test fixture first
     let messages = escargot::CargoBuild::new()
@@ -677,7 +232,7 @@ fn pcode_coverage() -> processor::Result<()> {
         .exec()
         .unwrap();
 
-    let image = match find_test_fixture(messages) {
+    let image = match common::find_test_fixture(messages) {
         Ok(image) => image,
         Err(messages) => {
             panic!("Failed to find test fixture: {messages:?}");
@@ -686,21 +241,21 @@ fn pcode_coverage() -> processor::Result<()> {
 
     let sleigh = x86_64_sleigh().expect("failed to build sleigh");
     let rip = sleigh.register_from_name("RIP").expect("invalid register");
-    let mut memory = memory_with_image(&sleigh, image, &rip);
-    init_registers_x86_64(&sleigh, &mut memory);
+    let mut memory = common::memory_with_image(&sleigh, image, &rip);
+    common::init_registers_x86_64(&sleigh, &mut memory);
 
     // Init stack address in memory to magic EXIT_IP_ADDR value
     let stack_addr = VarnodeData {
         address: Address {
-            offset: INITIAL_STACK,
+            offset: common::INITIAL_STACK,
             address_space: sleigh
                 .address_space_by_name("ram")
                 .expect("failed to find ram"),
         },
-        size: EXIT_IP_ADDR.to_le_bytes().len(),
+        size: common::EXIT_IP_ADDR.to_le_bytes().len(),
     };
     memory
-        .write(&stack_addr, EXIT_IP_ADDR.into())
+        .write(&stack_addr, common::EXIT_IP_ADDR.into())
         .expect("failed to initialize stack");
 
     let handler = ProcessorHandlerX86::new(&sleigh);
@@ -731,7 +286,8 @@ fn pcode_coverage() -> processor::Result<()> {
             .try_into()
             .expect("failed to concretize RIP");
 
-        if instruction_pointer == EXIT_IP_ADDR && matches!(processor.state(), ProcessorState::Fetch)
+        if instruction_pointer == common::EXIT_IP_ADDR
+            && matches!(processor.state(), ProcessorState::Fetch)
         {
             break;
         }
@@ -781,7 +337,7 @@ fn pcode_coverage_aarch64() -> processor::Result<()> {
         .exec()
         .unwrap();
 
-    let image = match find_test_fixture(messages) {
+    let image = match common::find_test_fixture(messages) {
         Ok(image) => image,
         Err(messages) => {
             panic!("Failed to find test fixture: {messages:?}");
@@ -790,21 +346,21 @@ fn pcode_coverage_aarch64() -> processor::Result<()> {
 
     let sleigh = common::aarch64_sleigh().expect("failed to build sleigh");
     let pc = sleigh.register_from_name("pc").expect("invalid register");
-    let mut memory = memory_with_image(&sleigh, image, &pc);
-    init_registers_aarch64(&sleigh, &mut memory);
+    let mut memory = common::memory_with_image(&sleigh, image, &pc);
+    common::init_registers_aarch64(&sleigh, &mut memory);
 
     // Init stack address in memory to magic EXIT_IP_ADDR value
     let stack_addr = VarnodeData {
         address: Address {
-            offset: INITIAL_STACK,
+            offset: common::INITIAL_STACK,
             address_space: sleigh
                 .address_space_by_name("ram")
                 .expect("failed to find ram"),
         },
-        size: EXIT_IP_ADDR.to_le_bytes().len(),
+        size: common::EXIT_IP_ADDR.to_le_bytes().len(),
     };
     memory
-        .write(&stack_addr, EXIT_IP_ADDR.into())
+        .write(&stack_addr, common::EXIT_IP_ADDR.into())
         .expect("failed to initialize stack");
 
     let handler = arch::aarch64::processor::ProcessorHandler::new(&sleigh);
@@ -835,7 +391,8 @@ fn pcode_coverage_aarch64() -> processor::Result<()> {
             .try_into()
             .expect("failed to concretize pc register");
 
-        if instruction_pointer == EXIT_IP_ADDR && matches!(processor.state(), ProcessorState::Fetch)
+        if instruction_pointer == common::EXIT_IP_ADDR
+            && matches!(processor.state(), ProcessorState::Fetch)
         {
             break;
         }
@@ -917,8 +474,8 @@ fn z3_integration() -> processor::Result<()> {
     write_register(&mut memory, "RIP", &code_offset.to_le_bytes());
 
     // Initialize stack and base pointer registers
-    write_register(&mut memory, "RSP", &INITIAL_STACK.to_le_bytes());
-    write_register(&mut memory, "RBP", &INITIAL_STACK.to_le_bytes());
+    write_register(&mut memory, "RSP", &common::INITIAL_STACK.to_le_bytes());
+    write_register(&mut memory, "RBP", &common::INITIAL_STACK.to_le_bytes());
 
     // Put EXIT_IP_ADDR onto the stack. The final RET will trigger this.
     memory.write(
@@ -927,11 +484,11 @@ fn z3_integration() -> processor::Result<()> {
                 sleigh
                     .address_space_by_name("ram")
                     .expect("failed to find ram"),
-                INITIAL_STACK,
+                common::INITIAL_STACK,
             ),
-            EXIT_IP_ADDR.to_le_bytes().len(),
+            common::EXIT_IP_ADDR.to_le_bytes().len(),
         ),
-        EXIT_IP_ADDR.into(),
+        common::EXIT_IP_ADDR.into(),
     )?;
 
     let input_value = sym::SymbolicBitVec::with_size(32);
@@ -968,7 +525,7 @@ fn z3_integration() -> processor::Result<()> {
                 .expect("failed to read RIP")
                 .try_into()
                 .expect("failed to concretize RIP");
-            if rip_value == EXIT_IP_ADDR {
+            if rip_value == common::EXIT_IP_ADDR {
                 finished.push(processors.swap_remove(i));
 
                 // Do not update index since processor from end was moved here
@@ -1130,8 +687,8 @@ fn take_the_path_not_taken() -> processor::Result<()> {
     write_register(&mut memory, "RIP", &code_offset.to_le_bytes());
 
     // Initialize stack and base pointer registers
-    write_register(&mut memory, "RSP", &INITIAL_STACK.to_le_bytes());
-    write_register(&mut memory, "RBP", &INITIAL_STACK.to_le_bytes());
+    write_register(&mut memory, "RSP", &common::INITIAL_STACK.to_le_bytes());
+    write_register(&mut memory, "RBP", &common::INITIAL_STACK.to_le_bytes());
 
     // Put EXIT_IP_ADDR onto the stack. The final RET will trigger this.
     memory.write(
@@ -1140,11 +697,11 @@ fn take_the_path_not_taken() -> processor::Result<()> {
                 sleigh
                     .address_space_by_name("ram")
                     .expect("failed to find ram"),
-                INITIAL_STACK,
+                common::INITIAL_STACK,
             ),
-            EXIT_IP_ADDR.to_le_bytes().len(),
+            common::EXIT_IP_ADDR.to_le_bytes().len(),
         ),
-        EXIT_IP_ADDR.into(),
+        common::EXIT_IP_ADDR.into(),
     )?;
 
     // Create symbolic input
@@ -1200,7 +757,8 @@ fn take_the_path_not_taken() -> processor::Result<()> {
             .read(&rip)?
             .try_into()
             .expect("failed to concretize RIP");
-        if instruction_pointer == EXIT_IP_ADDR && matches!(processor.state(), ProcessorState::Fetch)
+        if instruction_pointer == common::EXIT_IP_ADDR
+            && matches!(processor.state(), ProcessorState::Fetch)
         {
             break;
         }
